refactor(db): db-dialect a Postgres-only y cierra gaps en SQL crudo
- db-dialect.ts: elimina las ramas SQLite y detectDialect(); mantiene la
  API `sql.*` (timestamp, dateCol, bool, nowMinusDays, now, monthCol) para
  no reescribir los 6 archivos que la usan. Quita tsGte/tsLte (sin uso).
- super-admin/statistics: literales `'${date.toISOString()}'` -> `sql.timestamp(date)`,
  consistente con el resto de queries; comentario de topStores actualizado
  (ya no hay SQLite).
- reports/informes (inv-valuation): `::float8` en los SUM de Decimal para
  devolver number en vez de string/Decimal ambiguo.

Verificado contra el Postgres local: ambas queries ejecutan OK.

+++ src/app/api/reports/informes/route.ts
@@ -65,8 +65,8 @@ export async function GET(request: NextRequest) {
 
       // 2. INVENTARIO — Valuación y días
       safe('inv-valuation', () => db.$queryRawUnsafe(`
-        SELECT COALESCE(SUM(cost_price * current_stock), 0) as "totalCost",
-               COALESCE(SUM("salePrice" * current_stock), 0) as "totalRetail",
+        SELECT COALESCE(SUM(cost_price * current_stock), 0)::float8 as "totalCost",
+               COALESCE(SUM("salePrice" * current_stock), 0)::float8 as "totalRetail",
                COUNT(CASE WHEN current_stock = 0 THEN 1 END) as "outOfStock",
                COUNT(CASE WHEN current_stock <= min_stock THEN 1 END) as "lowStock",
                COUNT(*) as "totalProducts"
